pyproject.toml
[project]
name = "sample-app"
version = "0.1.0"
description = "Sample app with Strands agent and MLflow evaluation"
requires-python = ">=3.13"
dependencies = [
#    "mlflow @ git+https://github.com/LeviLong01/mlflow.git@genai-otel-semantic-conventions",
    "mlflow==3.10.1",
    "strands-agents==1.29.0",
    "strands-agents-tools==0.2.22",
    "anthropic>=0.79.0",
    "python-dotenv",
    "pandas",
    "numpy",
    "openai",
    "litellm",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
include = ["*.py"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
